Getting a framework going
So the site works pretty well on my xampp server but the pages dont load
under the navbar in the pi. I made an index.php which loads the sub
pages based on get. I fowarded my xampp server so that you can connect
to it. Its the same ip as the pi but port 220.
http://72.204.14.125:220/
I also worked on the mysql driver whilst  remaking the
login
registration
etc.. pages