Event form that sends stuff to mysql
i made an event form in the modal and it sends its data to mysql. ill
work on cleaning it up and trying to populate the various calendar
views with the events. right now it just sends a 1 for the group id,
once we get a better hold on groups, ill revise that too.