Nav bar dropdown
Pretty basic navbar dropdown menu. Maybe I'm missing something, but the
dropdown functionality and links only work on the "today" page for me
for some reason.

There are a few ways this could be formatted. One I thought of
separating the username and arrow so you could click the username to go
home or click the arrow for the dropdown menu. Will work on it a bit
more.