Help / Navbar / Event work
Filled in some questions on the help page to fill it out.

Navbar dropdown menu is basically done. Just have to figure out why the
dropdown won't work on pages other than "Today"

Gonna start working on trying to push event settings to the database
when modified on the event page